Add task to scan generated code to make sure versions match
